# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List non-conformity reports
      description: Returns a filtered, paginated list of non-conformity reports
      parameters:
        - name: defectType
          in: query
          required: false
          schema:
            type: string
        - name: syncStatus
          in: query
          required: false
          schema:
            type: string
        - name: dateFrom
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: dateTo
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated list of reports
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportListResponse"
    post:
      operationId: createReport
      tags: [reports]
      summary: Create a non-conformity report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReportBody"
      responses:
        "201":
          description: Report created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /reports/{id}:
    put:
      operationId: updateReport
      tags: [reports]
      summary: Update a non-conformity report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReportBody"
      responses:
        "200":
          description: Report updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Report is locked (SLA expired)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Report not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      operationId: deleteReport
      tags: [reports]
      summary: Delete a non-conformity report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Report deleted
        "404":
          description: Report not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    get:
      operationId: getReport
      tags: [reports]
      summary: Get a single report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Report found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "404":
          description: Report not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /reports/{id}/sync-status:
    patch:
      operationId: updateReportSyncStatus
      tags: [rpa]
      summary: Update sync status of a report
      description: Used by the RPA bot to update the sync status after processing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSyncStatusBody"
      responses:
        "200":
          description: Sync status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "400":
          description: Invalid status value
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Report not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /reports/{id}/qc:
    put:
      operationId: updateReportQc
      tags: [reports]
      summary: QC 분석 결과 저장 (admin 전용)
      description: QC 담당자가 현장 분석 후 조치 결과를 저장한다. isLocked 보고서도 저장 가능. 원본 보고서 일부 필드(itemCode, flawTypeCd, lostManHours)도 함께 수정 가능.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QcAnalysisBody"
      responses:
        "200":
          description: QC analysis saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Admin role required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Report not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /reports/{id}/status:
    patch:
      operationId: updateReportStatus
      tags: [reports]
      summary: QC 워크플로우 상태 전이
      description: |
        역할 기반 상태 전이 매트릭스를 적용한다.
        - reviewer: OPEN→IN_REVIEW, IN_REVIEW→PENDING_COLLAB, PENDING_COLLAB→RESOLVED
        - approver: RESOLVED→APPROVED
        - admin: 모든 전이 허용
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStatusBody"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "400":
          description: Invalid transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Role not permitted for this transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Report not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/close-month:
    post:
      operationId: closeMonth
      tags: [reports]
      summary: 월 마감 (관리자 전용)
      description: 지정한 연월 이전 보고서를 일괄 isLocked=true 처리한다. 관리자만 호출 가능.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CloseMonthBody"
      responses:
        "200":
          description: 마감 처리 결과
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CloseMonthResult"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Admin role required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  schemas:
...
    ReportListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Report"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
      required:
        - data
        - total
        - page
        - pageSize
    CreateReportBody:
      type: object
      properties:
        reportDate:
          type: string
          format: date-time
        itemCode:
          type: string
        modelName:
          type: ["string", "null"]
        processName:
          type: string
        defectType:
          type: string
        description:
          type: string
        imageUrl:
          type: ["string", "null"]
        registrantName:
          type: ["string", "null"]
        ncrType:
          type: ["string", "null"]
        factory:
          type: ["string", "null"]
        shipmentUnit:
          type: ["string", "null"]
        lostManHours:
          type: ["number", "null"]
        defectQty:
          type: ["integer", "null"]
        occurrenceDate:
          type: ["string", "null"]
          format: date-time
        issuingTeam:
          type: ["string", "null"]
        plantCd:
          type: ["string", "null"]
        processCd:
          type: ["string", "null"]
        flawTypeCd:
          type: ["string", "null"]
        deptCd:
          type: ["string", "null"]
        ncrGbnCd:
          type: ["string", "null"]
        vendorCd:
          type: ["string", "null"]
          description: "거래처 코드 또는 사업자번호 (RPA가 ERP #11에 입력)"
        vendorNm:
          type: ["string", "null"]
          description: "거래처명"
        productType:
          type: string
          enum: ["양산", "개발"]
          description: "양산 or 개발"
        actionDirection:
          type: string
          description: "조치 방향"
          enum: ["업체 방문 수정", "생산팀 자체 수정", "업체 반출 및 수정 입고"]
        remarks:
          type: ["string", "null"]
          description: "비고"
        shipmentDateFrom:
          type: ["string", "null"]
          format: date-time
          description: "출하 기간 시작일"
        shipmentDateTo:
          type: ["string", "null"]
          format: date-time
          description: "출하 기간 종료일"
        managerCd:
          type: ["string", "null"]
          description: "담당자 코드"
        managerNm:
          type: ["string", "null"]
          description: "담당자명"
      required:
        - itemCode
        - processName
        - defectType
        - description
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...